test(taps): fix review findings on derive_offsets property test
The property test drew g0/g1 independently from the same 32-step grid;
adjacent steps are ~21% apart, an order of magnitude outside the 2%
tolerance band, so the uniformity guard only ever fired on the diagonal
(32 of 1024 pairs). Off the diagonal the assertion restated the guard's
own arithmetic against itself and could not fail for any change to it.

Adds a second grid that perturbs each g0 by ratios straddling the guard's
real decision boundary (derived here as r in [0.9608, 1.0408] for
kUniformTol=0.02, matching the review's claim) and asserts both the
property (never uniform) and the guard's actual decision, pinned to
literal documented constants rather than the live tap_tuning:: values so
the assertion can't become tautological against a change to either one --
confirmed live: an earlier version keyed to the live kUniformSpread did
not catch a kUniformSpread mutation, and was corrected before committing.

Also: the property test's `checked > 500` vacuity guard is replaced with
an exact count (the grid can never reach the mute paths, so `> 500` proved
nothing); a case for the untested post-spread kMinGap bail-out
(view(40,40)) is added; two tolerance-band tests gained an absolute-offset
CHECK alongside their existing spacing CHECK; and derive_offsets' o0+o1
sum is widened to int64_t to close a signed-overflow UB path on
unconstrained int32_t inputs.

Minor 7 (the claim that the grid's gap values are unreachable due to
tick() quantisation) was overruled by the review itself and not
implemented.

+++ engine/fx/taps.cpp
@@ -31,9 +31,13 @@ void spky::derive_offsets(const RhythmView& rv, int32_t tape_len, int32_t out[2]
 
     const int32_t limit = tape_len - 2;
     const int32_t o0 = g0;
-    const int32_t o1 = g0 + g1;
+    // int64_t: g0 and g1 are otherwise-unconstrained int32_t, so g0 + g1 can
+    // overflow int32_t (signed overflow is UB) for large inputs. Widening the
+    // sum keeps this comparison exact without narrowing before the bound
+    // check below, which is what actually decides whether o1 is representable.
+    const int64_t o1 = static_cast<int64_t>(g0) + static_cast<int64_t>(g1);
     // Mute, never clamp: clamping would put two taps at the same position,
     // turning a missing echo into a doubled one.
     if (o0 <= limit) out[0] = o0;
-    if (o1 <= limit) out[1] = o1;
+    if (o1 <= limit) out[1] = static_cast<int32_t>(o1);
 }
